% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_server.R
\name{classification_server}
\alias{classification_server}
\alias{innerServer_7}
\title{SVMDO}
\usage{
innerServer_7(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of wilks lambda filtration and SVM classification of disease filtered differentially expressed gene set
}
\description{
SVMDO
}
