% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_ui.R
\name{classification_ui}
\alias{classification_ui}
\alias{innerUI_classification}
\title{SVMDO}
\usage{
innerUI_classification(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of wilk's lambda filtration and SVM classification of disease filtered differentially expressed gene set
}
\description{
SVMDO
}
