% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_data_input_server.R
\name{clinic_data_input_server}
\alias{clinic_data_input_server}
\alias{innerServer_clinic}
\title{SVMDO}
\usage{
innerServer_clinic(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of loading clinical data
}
\description{
SVMDO
}
