% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_data_input_ui.R
\name{clinic_data_input_ui}
\alias{clinic_data_input_ui}
\alias{innerUI_clinic_data}
\title{SVMDO}
\usage{
innerUI_clinic_data(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of loading clinical data
}
\description{
SVMDO
}
