% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_server.R
\name{deg_server}
\alias{deg_server}
\alias{innerServer_3}
\title{SVMDO}
\usage{
innerServer_3(input, output, session, rawData, rval)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{rawData}{expression dataset provided from innerServer_exp_server}

\item{rval}{Selected radio button information provided from innerServer_rad_server}
}
\value{
Server section of differential gene expression analysis
}
\description{
SVMDO
}
