% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_ui.R
\name{deg_ui}
\alias{deg_ui}
\alias{innerUI_deg_analysis}
\title{SVMDO}
\usage{
innerUI_deg_analysis(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of differential gene expression analysis
}
\description{
SVMDO
}
