% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_selection_server.R
\name{directory_selection_server}
\alias{directory_selection_server}
\alias{innerServer}
\title{SVMDO}
\usage{
innerServer(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of entering output/working directory
}
\description{
SVMDO
}
