% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_gene_dw_ui.R
\name{disc_gene_download_ui}
\alias{disc_gene_download_ui}
\title{SVMDO}
\usage{
disc_gene_download_ui(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of final discriminative gene set download button
}
\description{
SVMDO
}
