% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_gene_dw_server.R
\name{disc_gene_dw_server}
\alias{disc_gene_dw_server}
\title{SVMDO}
\usage{
disc_gene_dw_server(input, output, session, gene_list_val, global)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of final discriminative gene set download button
}
\description{
SVMDO
}
