% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_based_gene_filtration_server.R
\name{do_based_gene_filtration_server}
\alias{do_based_gene_filtration_server}
\alias{innerServer_6}
\title{SVMDO}
\usage{
innerServer_6(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of disease ontology based filtration of differentially expressed genes
}
\description{
SVMDO
}
