% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_based_gene_filtration_ui.R
\name{do_based_gene_filtration_ui}
\alias{do_based_gene_filtration_ui}
\alias{innerUI_disease_ont_class}
\title{SVMDO}
\usage{
innerUI_disease_ont_class(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of disease ontology based filtration of differentially expressed genes
}
\description{
SVMDO
}
