% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_dataset_input_server.R
\name{expression_dataset_input_server}
\alias{expression_dataset_input_server}
\alias{innerServer_exp}
\title{SVMDO}
\usage{
innerServer_exp(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of providing expression dataset
}
\description{
SVMDO
}
