% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_directory_selection_ui.R
\name{gene_directory_selection_ui}
\alias{gene_directory_selection_ui}
\alias{innerUI_path}
\title{SVMDO}
\usage{
innerUI_path(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of entering output/working for gene list directory
}
\description{
SVMDO
}
