% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_name_server.R
\name{gene_list_name_server}
\alias{gene_list_name_server}
\alias{innerServer_10}
\title{SVMDO}
\usage{
innerServer_10(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of entering final gene list name
}
\description{
SVMDO
}
