% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_name_ui.R
\name{gene_list_name_ui}
\alias{gene_list_name_ui}
\alias{innerUI_gene_names}
\title{SVMDO}
\usage{
innerUI_gene_names(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of entering top gene value
}
\description{
SVMDO
}
