% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_obj_removal_server.R
\name{gui_obj_removal_server}
\alias{gui_obj_removal_server}
\alias{innerServer_9}
\title{SVMDO}
\usage{
innerServer_9(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of workspace clearance
}
\description{
SVMDO
}
