% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_obj_removal_ui.R
\name{gui_obj_removal_ui}
\alias{gui_obj_removal_ui}
\alias{innerUI_clear_env}
\title{SVMDO}
\usage{
innerUI_clear_env(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of workspace clearance
}
\description{
SVMDO
}
