% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_dataset_input_ui.R
\name{innerServer_exp_ui}
\alias{innerServer_exp_ui}
\alias{innerUI_exp_data}
\title{SVMDO}
\usage{
innerUI_exp_data(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of providing expression dataset into GUI
}
\description{
SVMDO
}
