% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_list_server.R
\name{plot_list_server}
\alias{plot_list_server}
\title{SVMDO}
\usage{
plot_list_server(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of preparing plot list to be visualized in GUI page
}
\description{
SVMDO
}
