% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_list_ui.R
\name{plot_list_ui}
\alias{plot_list_ui}
\alias{innerUI_collect_plot_data}
\title{SVMDO}
\usage{
innerUI_collect_plot_data(id)
}
\arguments{
\item{id}{connection output}
}
\value{
UI section of preparing plot list to be visualized in GUI page
}
\description{
SVMDO
}
