% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_push_server.R
\name{plot_push_server}
\alias{plot_push_server}
\title{SVMDO}
\usage{
plot_push_server(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of providing information about total number of survival plots for visualization
}
\description{
SVMDO
}
