% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_push_ui.R
\name{plot_push_ui}
\alias{plot_push_ui}
\alias{innerUI_plot_inject}
\title{SVMDO}
\usage{
innerUI_plot_inject(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of providing information about total number of survival plots for visualization
}
\description{
SVMDO
}
