% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_show_server.R
\name{plot_show_server}
\alias{plot_show_server}
\title{SVMDO}
\usage{
plot_show_server(input, output, session, max_data)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{max_data}{Information of total number of survival plots prepared with discriminative gene set}
}
\value{
Server section of providing information about total number of survival plots for visualization
}
\description{
SVMDO
}
