% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runGUI}
\alias{runGUI}
\title{Initiating GUI screen}
\usage{
runGUI(reproducible = TRUE)
}
\arguments{
\item{reproducible}{Visualizing GUI screen}
}
\value{
Returning GUI window screen
}
\description{
Function of SVMDO initiating GUI screen of main dialog box including
analysis steps
}
\examples{
#SVMDO::runGUI() Calling GUI without activating library
#runGUI() Calling GUI after activating library
# Disease Ontology Enrichment of a differentially expresed gene (entrez id):
a_1<-DOSE::enrichDO(2981,ont="DO", maxGSSize=Inf)
}
