% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot_dw_server.R
\name{surv_plot_dw_server}
\alias{surv_plot_dw_server}
\title{SVMDO}
\usage{
surv_plot_dw_server(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of downloading survival plots of discriminative gene set
}
\description{
SVMDO
}
