% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot_dw_ui.R
\name{surv_plot_dw_ui}
\alias{surv_plot_dw_ui}
\alias{surv_plots_download_ui}
\title{SVMDO}
\usage{
surv_plots_download_ui(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of downloading survival plots of discriminative gene set
}
\description{
SVMDO
}
