% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis_server.R
\name{survival_analysis_server}
\alias{survival_analysis_server}
\alias{innerServer_8}
\title{SVMDO}
\usage{
innerServer_8(input, output, session, rawData_2, rval)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{rawData_2}{Clinical data provided from clinic_data_input_server}

\item{rval}{Selected radio button information provided from innerServer_rad_server}
}
\value{
Server section of survival analysis of final discriminative gene set
}
\description{
SVMDO
}
