% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis_ui.R
\name{survival_analysis_ui}
\alias{survival_analysis_ui}
\alias{innerUI_surv}
\title{SVMDO}
\usage{
innerUI_surv(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of survival analysis of final discriminative gene set
}
\description{
SVMDO
}
