% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_level_info.R
\docType{package}
\name{SVMDO}
\alias{SVMDO}
\title{SVMDO}
\section{Package Description}{

It is an easy-to-use GUI using disease information for detecting tumor/normal sample
discriminating gene sets from differentially expressed genes. Our approach is based on an
iterative algorithm filtering genes with disease ontology enrichment analysis and wilk’s lambda
criterion connected to SVM classification model construction. Along with gene set extraction,
SVMDO also provides individual prognostic marker detection among the discriminative genes. The algorithm is designed
for FPKM and RPKM normalized RNA-Seq transcriptome datasets.
}

\section{Package Sections}{

\enumerate{
\item \strong{Analysis:} Acquiring discriminative gene sets and further detecting the gene subset with prognostic characteristics
\item \strong{Result:} Visualization and Download of discriminative gene sets and survival plot list of prognostic genes
}
}

\section{Steps of Analysis Screen}{

\enumerate{
\item To search your transcriptome dataset, use the file detection in Choose Your Expression Dataset section. The file will be automatically uploaded into the GUI.
\item To prevent clashing with test datasets, select "None" option from the radio button section.
\item By clicking on DEG Analysis button you further apply differential expression analysis. Labels of tissue_type column in dataset must contain “Nor” and “Tum” for determining normal/tumour (or tumor) samples. A message window saying Process Completed will appear if there is not any problem.
\item When the differential expression process is completed, a user-defined input size (n) is selected to filter the initial gene list (i.e., n number of upregulated and downregulated genes) by entering a number in Input Size section. It is predetermined as 50 in GUI which can be changed based on the user. A message window saying process completed will appear if there is not any problem. If there is problem with the value of input size, you will get a warning about inappropriate input size selection. If the input size remains, algorithm selects all of the differentiallly expressed genes to be used in the next process.
\item To apply disease ontology-based gene filtration, click on DO Analysis button. A message window saying process completed will appear if there is not any problem.
\item To further apply the following feature selection and classification processes, click on the Classification button. A message window saying process completed will appear if there is not any problem.
\item Acquired discriminative gene set can be further used for survival analysis to detect individual prognostic genes. To apply this process, use the file detection in Choose Clinical Data section for searching clinical data about patient survival followed by clicking on Survival Analysis button. A message window saying process completed will appear if there is not any problem.
}
}

\section{Steps of Result Screen}{

\enumerate{
\item To visualize discriminative gene sets inside GUI screen, click on Show Gene Results button. When you click this button, a table of gene set will appear. If there is a problem in the analysis, an error message will appear.
\item To visualize survival plots of individual genes, two steps have to be applied. First of all, click on Prepare Plot Lists button to feed plot information to the visualization system. After that, click on Show Plots button to visualize survival plots. If there is a problem in the analysis, an error message will appear at each step.
\item Before downloading files, you can adjust the output directory with Choose Directory button. It can be used for separating files by selecting a destination before clicking download buttons. If it is desired, files can be downloaded to the same folder by selecting an output directory just one time before the download steps.If you do not select any output directory, files will be downloaded to your working directory.
\item To download the resulting discriminative gene set, it is obligatory to define a filename in the Enter Final Gene Set Filename section. After that, you can click on Download Gene List button to complete the process.
\item To download survival plots, you have to click on Download Plot List button. Names of plot files are automatically done by assigning gene names.
}
}

