% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_server.R
\name{table_server}
\alias{table_server}
\title{SVMDO}
\usage{
table_server(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of providing discriminative gene set for preparing table
}
\description{
SVMDO
}
