% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_ui.R
\name{table_ui}
\alias{table_ui}
\alias{innerUI_table_show}
\title{SVMDO}
\usage{
innerUI_table_show(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of providing discriminative gene set for preparing table
}
\description{
SVMDO
}
