% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_selection_server.R
\name{test_data_selection_server}
\alias{test_data_selection_server}
\alias{innerServer_rad}
\title{SVMDO}
\usage{
innerServer_rad(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of providing information about selected radio button
}
\description{
SVMDO
}
