% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_selection_ui.R
\name{test_data_selection_ui}
\alias{test_data_selection_ui}
\alias{innerUI_test_data}
\title{SVMDO}
\usage{
innerUI_test_data(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of providing information about selected radio button
}
\description{
SVMDO
}
