% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_val_based_deg_filtration_server.R
\name{top_val_based_deg_filtration}
\alias{top_val_based_deg_filtration}
\alias{innerServer_5}
\title{SVMDO}
\usage{
innerServer_5(input, output, session, top_val)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{top_val}{top gene number value provided from top_val_server}
}
\value{
Server section of selecting differentially expressed genes based on top gene value
}
\description{
SVMDO
}
