% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_val_based_deg_filtration_ui.R
\name{top_val_based_deg_filtration_ui}
\alias{top_val_based_deg_filtration_ui}
\alias{innerUI_top_gene_selection}
\title{SVMDO}
\usage{
innerUI_top_gene_selection(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of selecting differentially expressed genes based on top gene value
}
\description{
SVMDO
}
