% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_val_server.R
\name{top_val_server}
\alias{top_val_server}
\alias{innerServer_4}
\title{SVMDO}
\usage{
innerServer_4(input, output, session)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}
}
\value{
Server section of entering top gene value
}
\description{
SVMDO
}
