% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_val_ui.R
\name{top_val_ui}
\alias{top_val_ui}
\alias{innerUI_top_gene_val}
\title{SVMDO}
\usage{
innerUI_top_gene_val(id)
}
\arguments{
\item{id}{connection input}
}
\value{
UI section of entering top gene value
}
\description{
SVMDO
}
