% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_gene_dw_server.R
\name{disc_gene_dw_server}
\alias{disc_gene_dw_server}
\title{SVMDO}
\usage{
disc_gene_dw_server(input, output, session, gene_list_val)
}
\arguments{
\item{input}{server input}

\item{output}{server output}

\item{session}{server session}

\item{gene_list_val}{discriminative gene set list variable}
}
\value{
Server section of discriminative gene set download button
}
\description{
SVMDO
}
