% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runGUI}
\alias{runGUI}
\alias{linebreaks}
\title{SVMDO}
\usage{
linebreaks(n)
}
\arguments{
\item{n}{linebreak function variable}
}
\value{
Returning GUI window screen
}
\description{
SVMDO
}
\examples{
#SVMDO::runGUI() Calling GUI without activating library
#runGUI() Calling GUI after activating library
# Disease Ontology Enrichment of a differentially expresed gene (entrez id):
a_1<-DOSE::enrichDO(2981,ont="HDO")
}
